+++ CHANGELOG.md
@@ -12,7 +12,11 @@ package updates, you can specify your package dependency using
 
 ## [Unreleased]
 
-*No changes yet.*
+### Additions
+
+- The `sortedOverlap(with: by:)` and `sortedOverlap(with:)` methods have been
+  added.  They report how much overlap two sorted sequences have, expressed by
+  the new `Inclusion` type.
 
 ---